ml_service/research/requirements.txt
pandas
matplotlib
psycopg2-binary
numpy
